move HintForm componentent into scenes/game
