+++ src/scenes/game/Game.tsx
@@ -29,7 +29,7 @@ import {
 } from "src/store/actions";
 import { SharedState } from "src/store/SharedState";
 import { GuessForm } from "../GuessForm";
-import { HintForm } from "../HintForm";
+import { HintForm } from "./HintForm";
 import { Chadburn } from "./Chadburn";
 import styles from "./Game.module.css";
 import { Spectrum } from "./Spectrum";
