+++ src/scenes/game/HintForm.tsx
